requirements.txt
botocore==1.24.6
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.0
Flask==2.0.3
google-auth==2.6.0
google-auth-oauthlib==0.5.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.0.1
pymongo-auth-aws==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
